fix: add error handling for patch feature

+++ app/src/main/java/com/daluwi/sc_news/features/patches/data/repository/PatchRepositoryImpl.kt
@@ -2,6 +2,8 @@ package com.daluwi.sc_news.features.patches.data.repository
 
 import android.util.Log
 import com.daluwi.sc_news.core.connectivity.NetworkChecker
+import com.daluwi.sc_news.core.error_handling.RepositoryError
+import com.daluwi.sc_news.core.error_handling.Result
 import com.daluwi.sc_news.features.patches.data.source.local.PatchDAO
 import com.daluwi.sc_news.features.patches.data.source.local.toDomain
 import com.daluwi.sc_news.features.patches.data.source.local.toEntity
@@ -16,7 +18,7 @@ class PatchRepositoryImpl(
     private val api: PatchApi,
     private val networkChecker: NetworkChecker
 ) : PatchRepository {
-    override suspend fun getPatches(): List<Patch> {
+    override suspend fun getPatches(): Result<List<Patch>, RepositoryError> {
         return if (networkChecker.isAvailable()) {
             getRemotePatches()
         } else {
@@ -24,28 +26,35 @@ class PatchRepositoryImpl(
         }
     }
 
-    override suspend fun getLocalPatches(): List<Patch> {
-        return dao.getPatches().map { it.toDomain() }
+    override suspend fun getLocalPatches(): Result<List<Patch>, RepositoryError> {
+        return try {
+            Log.i("LOCAL", "loaded successfully")
+            Result.Success(dao.getPatches().map { it.toDomain() })
+        } catch (e: Exception) {
+            Log.e("LOCAL", e.printStackTrace().toString())
+            Result.Error(RepositoryError.LOCAL_FAILED)
+        }
+
     }
 
-    override suspend fun getRemotePatches(): List<Patch> {
-        try {
+    override suspend fun getRemotePatches(): Result<List<Patch>, RepositoryError> {
+        if (!networkChecker.isAvailable()) return Result.Error(RepositoryError.NO_INTERNET)
+        return try {
             val patches = api.getPatches().map { it.toDomain() }
             dao.deleteAll()
             dao.insertAll(patches.map { it.toEntity() })
-            return patches
+            Log.i("REMOTE", "loaded successfully")
+            Result.Success(patches)
         } catch (e: Exception) {
-            // TODO Should implement a msg to show to the user
-            Log.d("FAILSAFE", "Error connecting to server. Using DB as fallback.")
-            Log.e("FAILSAFE", e.printStackTrace().toString())
-            return getLocalPatches()
+            Log.e("REMOTE", e.printStackTrace().toString())
+            Result.Error(RepositoryError.REMOTE_FAILED)
         }
     }
 
-
-    override suspend fun getPatchByChannel(channel: Channel): Patch? {
-        val entity = dao.getPatchByChannel(channel) ?: return null
-        return entity.toDomain()
+    override suspend fun getPatchByChannel(channel: Channel): Result<Patch, RepositoryError> {
+        val entity =
+            dao.getPatchByChannel(channel) ?: return Result.Error(RepositoryError.LOCAL_FAILED)
+        return Result.Success(entity.toDomain())
     }
 
     override suspend fun insertPatch(patch: Patch) {
